renaming encoding modules
