+++ sia/src/encoding.rs
@@ -1,8 +1,8 @@
-mod v1_encoding;
-mod v2_encoding;
+mod v1;
+mod v2;
 
 pub use sia_sdk_derive::{SiaDecode, SiaEncode};
-pub use v2_encoding::{Error, Result, SiaDecodable, SiaEncodable};
+pub use v2::{Error, Result, SiaDecodable, SiaEncodable};
 
 pub use sia_sdk_derive::{V1SiaDecode, V1SiaEncode};
-pub use v1_encoding::{V1SiaDecodable, V1SiaEncodable};
+pub use v1::{V1SiaDecodable, V1SiaEncodable};
